+++ app/routes/components/Tabs/changelog.codegen.ts
@@ -8,4 +8,5 @@ export const changelog = ''
   + '- 2025-11-15: Adjust spacing between tabs for better visual balance. Remove the\n'
   + '  dependency on deprecated `ThemeProvider` and use `CopyUiProvider` instead.\n'
   + '- 2025-12-07: Support the `size` prop.\n'
+  + '- 2025-12-13: Update to support the new color scheme.\n'
   ;