+++ src/components/Tabs/index.ts
@@ -1,3 +1,3 @@
-// Tabs from copy-ui @ 2025-12-07.
+// Tabs from copy-ui @ 2025-12-13.
 
 export { Tabs } from "./tabs";